feat: added wishlist reducer
