+++ src/context/AppContext.js
@@ -1,11 +1,19 @@
 import React from 'react'
 import SortedBrands from '../utils/SortedBrands'
-import filtersReducer, { data } from './filtersReducer'
+import filtersReducer from './filtersReducer'
+import wishListReducer from './wishListReducer'
+import { data } from './initialData'
+import combineReducers from 'react-combine-reducers';
 
 const AppContext = React.createContext(null)
 export function useAppContext() {
     return React.useContext(AppContext)
 }
+
+const [rootReducer, initialData] = combineReducers({
+    filters: [filtersReducer, data],
+    wishlist: [wishListReducer, []]
+});
 
 function AppCnxtProvider({ children }) {
     const [theme, setTheme] = React.useState('light')
@@ -14,7 +22,15 @@ function AppCnxtProvider({ children }) {
     };
 
     // reducer for filters and products
-    const [reducer, dispatch] = React.useReducer(filtersReducer, data)
+    const [state, dispatch] = React.useReducer(rootReducer, initialData)
+
+    const reducer = {
+        filters: state.filters.filters,
+        products: state.filters.products,
+        filteredProducts: state.filters.filteredProducts,
+        wishList: state.wishlist
+    }
+
     React.useEffect(() => {
         fetch("./MOCK_DATA.json")
             .then((res) => res.json())
