+++ src/context/filtersReducer.js
@@ -1,19 +1,4 @@
-export const data = {
-    filters: {
-        lowToHigh: false,
-        highToLow: false,
-        newestArrivals: false,
-        allPrice: true,
-        minPrice: null,
-        maxPrice: null,
-        rating: null,
-        brand: null,
-        payOnDelivery: false,
-        discount: null
-    },
-    products: [],
-    filteredProducts: []
-}
+import { data } from "./initialData";
 
 const filtersReducer = (state, action) => {
     switch (action.type) {
