+++ src/context/initialData.js
@@ -1,0 +1,17 @@
+export const data = {
+    filters: {
+        lowToHigh: false,
+        highToLow: false,
+        newestArrivals: false,
+        allPrice: true,
+        minPrice: null,
+        maxPrice: null,
+        rating: null,
+        brand: null,
+        payOnDelivery: false,
+        discount: null
+    },
+    products: [],
+    filteredProducts: [],
+    wishList: []
+}
